requirements.txt
beautifulsoup4==4.8.2
gensim==3.8.1
lxml==4.5.0
mistletoe==0.7.2
numpy==1.18.1
plac==1.1.3
scipy==1.4.1
spacy==2.2.3
tqdm==4.43.0
google-cloud-bigquery==1.24.0
google-cloud-bigquery-storage==0.8.0
plotly==4.14.3
pandas
click
jsonlines
